Add WorkerTarget and OtherTarget
